# backend/src/modules/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MikConnect API",
    "version": "1.0.0",
    "description": "API de gestion de hotspots WiFi, tickets, revendeurs et paiements pour MikConnect.",
    "contact": { "name": "MikConnect Support", "email": "support@mikconnect.com" }
  },
  "servers": [{ "url": "/api", "description": "API v1" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Voucher": { "type": "object", "properties": { "id": { "type": "string" }, "code": { "type": "string" }, "price": { "type": "number" }, "status": { "type": "string", "enum": ["ACTIVE", "USED", "EXPIRED"] }, "hotspotId": { "type": "string" }, "planId": { "type": "string" }, "expiresAt": { "type": "string", "format": "date-time" } } },
      "Transaction": { "type": "object", "properties": { "id": { "type": "string" }, "reference": { "type": "string" }, "amount": { "type": "number" }, "method": { "type": "string" }, "status": { "type": "string", "enum": ["PENDING", "COMPLETED", "FAILED"] }, "phoneNumber": { "type": "string" } } },
...
      "User": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "email": { "type": "string" }, "role": { "type": "string", "enum": ["ADMIN", "ISP", "RESELLER", "PARTNER"] }, "phone": { "type": "string" }, "country": { "type": "string" }, "active": { "type": "boolean" } } },
...
    }
  },
  "paths": {
...
    "/vouchers": {
      "get": {
        "tags": ["Vouchers"],
        "summary": "Lister les tickets",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Liste des tickets", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Voucher" } } } } } }
      },
      "post": {
        "tags": ["Vouchers"],
        "summary": "Generer des tickets",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "planId": { "type": "string" }, "hotspotId": { "type": "string" }, "count": { "type": "integer", "default": 1 } } } } } },
        "responses": { "201": { "description": "Tickets crees" } }
      }
    },
    "/vouchers/validate": {
      "post": {
        "tags": ["Vouchers"],
        "summary": "Valider un code",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "code": { "type": "string" }, "hotspotId": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Ticket valide" }, "404": { "description": "Code invalide ou expire" } }
      }
    },
    "/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Mes transactions",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Liste des transactions", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Transaction" } } } } } }
      }
    },
...
    "/users": {
      "get": {
        "tags": ["Utilisateurs"],
        "summary": "Lister les utilisateurs (admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Liste des utilisateurs", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/User" } } } } } }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["Utilisateurs"],
        "summary": "Mon profil",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Profil utilisateur", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } } } }
      },
      "put": {
        "tags": ["Utilisateurs"],
        "summary": "Mettre a jour mon profil",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "phone": { "type": "string" }, "country": { "type": "string" }, "companyName": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Profil mis a jour" } }
      }
    },
...
    "/mikrotik/test": {
      "post": {
        "tags": ["MikroTik"],
        "summary": "Tester la connexion a un routeur",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "host": { "type": "string" }, "port": { "type": "integer", "default": 8728 }, "username": { "type": "string" }, "password": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Resultat du test" } }
      }
    },
...
  },
...
}